+++ internal/backends/authority/organization.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/ugent-library/biblio-backoffice/internal/backends"
-	"github.com/ugent-library/biblio-backoffice/internal/backends/es6"
 	"github.com/ugent-library/biblio-backoffice/internal/models"
 )
 
@@ -31,18 +30,17 @@ type searchEnvelope struct {
 }
 
 func (c *Client) GetOrganization(id string) (*models.Organization, error) {
-
-	requestBody := es6.M{
-		"query": es6.M{
-			"term": es6.M{
+	requestBody := M{
+		"query": M{
+			"term": M{
 				"_id": id,
 			},
 		},
 		"size": 1,
 	}
-	var responseBody organizationSearchEnvelope = organizationSearchEnvelope{}
+	responseBody := organizationSearchEnvelope{}
 
-	if e := c.es.SearchWithBody("biblio_organization", requestBody, &responseBody); e != nil {
+	if e := c.search("biblio_organization", requestBody, &responseBody); e != nil {
 		return nil, e
 	}
 
@@ -57,7 +55,7 @@ func (c *Client) GetOrganization(id string) (*models.Organization, error) {
 }
 
 func (c *Client) SuggestOrganizations(q string) ([]models.Completion, error) {
-	limit := 500
+	limit := 20
 	completions := make([]models.Completion, 0, limit)
 
 	// remove duplicate spaces
@@ -67,11 +65,11 @@ func (c *Client) SuggestOrganizations(q string) ([]models.Completion, error) {
 	q = strings.TrimSpace(q)
 
 	qParts := strings.Split(q, " ")
-	queryMust := make([]es6.M, 0, len(qParts))
+	queryMust := make([]M, 0, len(qParts))
 
 	for _, qp := range qParts {
-		queryMust = append(queryMust, es6.M{
-			"query_string": es6.M{
+		queryMust = append(queryMust, M{
+			"query_string": M{
 				"default_operator": "AND",
 				"query":            fmt.Sprintf("%s*", qp),
 				"default_field":    "all",
@@ -80,9 +78,9 @@ func (c *Client) SuggestOrganizations(q string) ([]models.Completion, error) {
 		})
 	}
 
-	requestBody := es6.M{
-		"query": es6.M{
-			"bool": es6.M{
+	requestBody := M{
+		"query": M{
+			"bool": M{
 				"must": queryMust,
 			},
 		},
@@ -91,7 +89,7 @@ func (c *Client) SuggestOrganizations(q string) ([]models.Completion, error) {
 
 	var responseBody searchEnvelope = searchEnvelope{}
 
-	if e := c.es.SearchWithBody("biblio_organization", requestBody, &responseBody); e != nil {
+	if e := c.search("biblio_organization", requestBody, &responseBody); e != nil {
 		return nil, e
 	}