+++ internal/backends/authority/project.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/ugent-library/biblio-backoffice/internal/backends"
-	"github.com/ugent-library/biblio-backoffice/internal/backends/es6"
 	"github.com/ugent-library/biblio-backoffice/internal/models"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
@@ -74,18 +73,18 @@ func (c *Client) SuggestProjects(q string) ([]models.Completion, error) {
 	limit := 20
 	completions := make([]models.Completion, 0, limit)
 
-	var query es6.M = es6.M{
-		"match_all": es6.M{},
+	var query M = M{
+		"match_all": M{},
 	}
 
 	q = strings.TrimSpace(q)
 
 	if q != "" {
-		dismaxQueries := make([]es6.M, 0, len(projectFieldsBoosts))
+		dismaxQueries := make([]M, 0, len(projectFieldsBoosts))
 		for field, boost := range projectFieldsBoosts {
-			dismaxQuery := es6.M{
-				"match": es6.M{
-					field: es6.M{
+			dismaxQuery := M{
+				"match": M{
+					field: M{
 						"query":    q,
 						"operator": "AND",
 						"boost":    boost,
@@ -94,14 +93,14 @@ func (c *Client) SuggestProjects(q string) ([]models.Completion, error) {
 			}
 			dismaxQueries = append(dismaxQueries, dismaxQuery)
 		}
-		query = es6.M{
-			"dis_max": es6.M{
+		query = M{
+			"dis_max": M{
 				"queries": dismaxQueries,
 			},
 		}
 	}
 
-	requestBody := es6.M{
+	requestBody := M{
 		"query": query,
 		"size":  limit,
 		"sort":  []string{"_score:desc"},
@@ -109,7 +108,7 @@ func (c *Client) SuggestProjects(q string) ([]models.Completion, error) {
 
 	var responseBody searchEnvelope = searchEnvelope{}
 
-	if e := c.es.SearchWithBody("biblio_project", requestBody, &responseBody); e != nil {
+	if e := c.search("biblio_project", requestBody, &responseBody); e != nil {
 		return nil, e
 	}